Implement Makefile compatiability with TEE/NAAAIM.
Implementation of the C++ unit test driver for the trusted
execution version of the NAAAIM library suggested the need to
architect the Makefile for the TEE/fusion library in a similar
fashion.

This involved moving the test that adds the C++ based unit test
driver to after the specification of the CC variable.  The
C++TOOLS variable was completely removed as well.  This requires
that the clean target be run with the same compiler as was used
to generate the build.