Add target for binary installation.
This commit introduces the install-bin target which descends
through the source heirarchy and installs components needed to
implement the SRDE runtime.  These are the tools and files needed
to create launch tokens, load enclaves and conduct various unit
tests of the SRDE.

The DESTDIR variable is supported that allows the installation to
be rooted in a filesystem heirarchy of the developer's choice.

The plan will be to introduce an install-dev target that
installs the components needed to support SRDE based
applications.