Fix API break in IAS protocol.
This update introduces a fix for a break in the Intel Attestation
Services (IAS) API protocol that was introduced in a TCB roll-over
that was implemented on 07/12/2019.

During this roll-over enforcement was implemented to require that
the following header be included in the REST request used to
submit an SGX quote for attestation verification:

Content-Type: application/json

The v3 protocol implementation had been done and tested when the
IAS implementation was not rejecting requests that did not have
this header.  As a result the need for this header had not been
noted.

On 07/12/2019 the IAS implementation began enforcing a
requirement that this header be present.  Absence of the header
caused the following error message to be returned:

HTTP/1.1 400 Bad Request

Multiple exchanges with Intel finally confirmed the source of
this regression and the fix that would be required.

In the process it was noted that Intel had moved forward
with a shared secret authentication model that will replace the
certificate/key authentication model currently being used in
the production IAS service.  Support will subsequently be added for
this model when it becomes clear what the timeframe for migration
to the new authentication model is.