Enforce authentication of IAS server.
Verification of the authenticity of the IAS server was not
previously being conducted.  This update requires that wget
verify the authenticity of the IAS server when submitting
a quote for verification.

The following location is used as the source for the certificate
chain used for the authentication:

/opt/IDfusion/etc/ca-bundle.pem

The install-bin target was updated to install this certificate
bundle from the /distrib/SRDE distribution file.