Configure remote attestatin test to use default locations.
The test-remote-target utility that exercises support for
generation of an IAS based attestation report was not using
default locations for its enclave or launch token.  This made
running it as part of the standard unit test package problematic.

This update configures the utility to use the compiled in default
for the enclave location source path that is the current
directory if ENCLAVE_DIR is not specified at compilation time.

The default launch token will be in the following location:

/var/lib/IDfusion/tokens/LocalTarget.token

The -t command-line switch was added to the utility in order to
aid in-tree testing of the utility.  The -t switch specifies the
token location.