Rename SGX runtime support library.
In order to avoid any potential for trademark issues all
references to SGX will be removed from the SRDE development
resources.

The first step of this process is to rename the library that
provides runtime services from:

libSGXrdk.a

To:

libSRDEruntime.a

The build process was updated to reflect this change.