Convert enclave management object to new name.
This rather major update converts the basic HurdLib object that
manages an enclave from the following object name:

SGXenclave

To the following:

SRDEenclave

Consistent with our work to remove SGX references.

This is the most far reaching change since any application that
deals with an enclave does so through this object.