Fix descending into SRDE directory.
The update that changed the former SGX directory to the SRDE
directory missed a needed change to the Makefile that would
generate the recursion into the directory.  As a result the SRDE
directory would not be built which caused a dependency problem
when the ISOidentity directory was built.

This update modifies the Makefile to have a target for the SRDE
directory which causes the directory to be properly descended.