Implement development installation target.
This update implements the install-dev: target in relevant
directories.  The purpose of this target is to install the
development infrastructure needed to create SRDE based
applications.

Each of the major development libraries has its include files
install in in a sub-directory of /opt/IDfusion/include.  Since
each library is independently named they are installed in
/opt/IDfusion/lib.

Since the HurdLib package does not have an installation target
the top-level Makefile has explicit specifications for the
include files that are to be included as well as the library
file.