Remove loader heuristic.
Testing of the SRDE development environment on RHEL7 indicated
that the loader on the platform was generating an instance of a
segment with a non-zero offset that was conflicting with
heuristics that have been previously implemented in the loader.

Secondary to an attempt to address these edge cases a heuristic
was applied that automatically decremented the number of zero
pages to be added to a segment if memory size of the segment was
less then the amount loaded.  This may have worked at one time
but is currently generating an incorrect load with an enclave
linked by the RHEL7 loader.

Testing against all of the current SRDE enclaves plus the Intel
production enclaves suggests that the loader is working in this
environment with the heuristic removed.