Add initial implementation of sample code.
This update implements a directory named Sample in the SRDE
master directory.  The purpose of this directoy is to contain
code that implements sample SGX applications based on the SRDE.

These applications are installed in the following directory when
the install-dev target is invoked:

/opt/IDfusion/share

With the initial applications being deployed into the unit-tests
directory in the /opt/IDfusion/share directory.

The initial sample code implements the test-fusion utility and
enclave used to unit test the SRDEfusion library.