Directly link IDfusion supplied libraries.
Testing of the SRDE development environment indicated operational
problems with the miscellaneous libraries supplied for
compatibility with a native GCC compile.  These libraries are
located in the following directory:

/usr/local/IDfusion/lib

The previous strategy had been to simply reference these
libraries with a '-L DIRNAME' directive.  The SRDE_libs
distribution RPM included a file in the /etc/ld.so.conf.d
directory that specified that /usr/local/IDfusion/lib should be
included in the search path for the dynamic linker.

Unfortunately this created problems on Ubuntu which is using
version 1.1 of OpenSSL for its standard implementation.  The
shared library directory reference of the dynamic linker caused
the IDfusion supplied library to be used for any Ubuntu
applications which had linked against OpenSSL.

Due to some subtle difference between the Ubuntu implementation
of the OpenSSL library and the IDfusion supplied version binaries
that were linked against the Ubuntu implementation failed to
function when they were referencing the IDfusion supplied
implementation.

This issue was addressed by embedding the location of the
IDfusion supplied libraries in the SRDE executables by specifying
the location through the following linker directive:

-Wl,-rpath=/usr/local/IDfusion/lib

This commit applies this linker directive to all of the
Makefiles.

The resulting binaries were tested and confirmed to work on both
RHEL and Ubuntu implementations.