Adjust loader strategy again.
The following commit:

4577538d55626672a271c168fb900bfebddd4b8a
Remove loader heuristic.

Ended up breaking the functionality of the enclave loader and
began causing all of the Intel supplied enclaves to fail their
initialization.

After an extensive round of testing and development a new
strategy was developed for calculating the number of null pages
that need to be added when the memory size is larger then the
file size of a segment.  The new strategy is the simplest of the
approaches that has been developed to date.

If the size of the segment loaded minus the offset of the segment
is greater then or equal to the virtual memory size the number of
additional pages is set to zero.

If this condition failes the size of the zero page segement is
the virtual memory size minus the current loaded size of the
segment minus the offset, ie:

size = segment->phdr.p_memsz - (loaded - offset)

This change causes all of the Intel production enclaves to
successfully load in addition to all the SRDE supplied enclaves.
This new strategy also allows the test-fusion unit test enclave
that was compiled under RHEL7 to properly initialize.  A
regression with the initialization of this enclave caused the
back referenced commit to be considered.