Rename SGXrsa object.
This commit renames the following object:

SGXrsa

To the following:

SRDErsa

In order to continue the removal of references to the SGX phrase.

The SRDErsa object is used exclusively by the EPID provisioning
code which has to encrypt/sign using an RSA key that uses SHA256
based Optimal Asymmetric Encryption Padding (OAEP).  The
functionality in this object arguably needs to transition to the
RSAkey object in the NAAAIM library.