Rename exception handling infrastructure.
This update converts the name of the file that implements the
enclave exception handling code from sgx_exception.c to
srde-exception.c.

The 'sgx' keyword was dropped from the functions that are
implemented to handle exception processing.  The function name
for installing an enclave specific exception handler was changed
to srde_configure_exception to link it to the SRDE.