Remove reference to sgx in OCALL Duct object manager.
The code that ties an enclave based Duct object(s) to untrusted
implementations of the object was in a file that prior to this
update was named Duct_sgxmgr.  This file was renamed to Duct_mgr
to remove the sgx reference.

The dispatch call has to be referenced in the OCALL function
dispatch array that is passed to an enclave implementing Duct
objects.  This external function was renamed and its use
instances were also changed correspondingly.