Use position independent code in libraries.
The linker on Ubuntu 18 is generating position independent
executables by default.  In addition the linker is verifying that
all of the referenced object files are compiled as position
independent code.  If not the linkage is terminated with an
error.

Complicating the matter is that the Ubuntu C chain is not
respecting the -fno-pie command-line option to disable the
generation of position independent executables.  Instead the
-no-pie command-line option is used to disable position
independent executables.

Since it is the path forward all of the library code was modified
to use the -fpic command-line option to generate code that was
compatible with position independent executables on Ubuntu 18.