Implement key generation in RSAkey.
This update adds the following three methods to the object that
manages RSA keys:

RSAkey->generate_key

RSAkey->get_public_key

RSAkey->get_private_key

The first function generates an RSA key with a specified size in
bits.

The two remaining methods return the PEM encoded form of either
the public or the private key.

The RSAkey unit test was updated to recognize the keyword
'generate' on the command-line.  This causes a key of size 2048
to be generated.  The components of the key are printed out along
with the PEM encoded version of the key.