Implement dynamic buffer sizing for trusted fprintf.
The strategy within the SRDE is to implement the printf, fprintf,
fputs and fputc functions as calls to a function that writes the
formatted output into a buffer.  That buffer is then passed to
untrusted space to be output.  The fputs and fputc are internally
mapped into a call to fprintf.

The previous implementation used a 1024 byte buffer that was
dynamically allocated from the stack in the fprintf
implementation.  The 1024 byte allocation ended up being
restrictive for some appliations that output a large amount of
text in a single call.

Since the fast majority of I/O is going to be far smaller then
1,024 bytes it was deemed inappropriate to increase the size of
the stack based buffer so a bypass scheme was implemented.

The vsnprintf function was used in 'bypass' mode where it returns
the size of the fully formatted string if a NULL pointer is
passed as the output buffer specifier.  If the desired output is
greater then 1,204 a buffer is dynamically allocated with malloc
and used as the output buffer.