Implement SRDEocall object and dynamic OCALL dispatch tables.
This update implements the SRDEocall object that implements the
ability to dynamically create the function table that defines the
functions in untrusted space that are to be called in response to
invoking a specific OCALL slot number from within an enclave.

Previous to the implementation of this object the OCALL function
table was statically created by defining the OCALL_api structure
as a constant structure in the outer scope of an application that
was calling an enclave.  The array of function pointers that
implement the dispatch table was represented as a variable length
array at the end of the OCALL_api structure.

In addition to being error prone and requiring boilerplate in
every application the variable length array was problematic for
C++ compilation.

The SRDEocall object allows a list of functions that will be
invoked in response to OCALL's to be built dynamically.  The
following method is used to add a single function:

->add

Once the object contains all of the necessary functions the
following method is used create a structure within a Buffer
object that represents the contents of an OCALL_api dynamic
array:

->get_table

This method accepts a pointer to a pointer to an OCALL_api
structure.  The overlay pointer can then be passed directory to
the ->boot_slot method of an SRDEenclave object.

The objective of this infrastructure is to allow the dispatch
table to be built dynamically by multiple independent library
systems.  Subsequent patches will implement this infrastructure
for the SRDEfusion and SRDEnaaaim libraries.