Consolidate SRDEfusion OCALL management.
This commit consolidates the OCALL implementation functions into
the SRDEruntime library.  This provides a framework for removing
all of the individual functions in enclave based applications in
place of using a single standardized implementation.

The OCALL definitions and implementations for the SRDEfusion
trusted library arein the following files:

SRDE/SRDEfusion-ocall.h

SRDE/SRDEfusion-ocall.c

The SRDEfusion-ocall.h file is designed to be included by the
SRDEfusion library so as to maintain consistency between
untrusted and enclave implementations.

All of the OCALL functions are declared with static scope in the
SRDEfusion-ocall.c file.  An external function is provided that
takes an SRDEocall object as an arguement and registers all of
the statically scoped functions.

The SRDEocall object was modified to add the following method:

->add_SRDEfusion

This method uses the SRDEfusion registration function to add all
of the defined functions to the object.  Calling this method
results in the creation of all the needed OCALL infrastructure
for applications that use the SRDEfusion library.