Fix Makefile for SRDEnaaaim library and unit tests.
Perhaps secondary to work on the dynamic OCALL manager
linking of the test-naaaim utility began to fail.  This was
addressed by adding the a specification for the SSL libcrypto
library.  It is unclear why this was working previously.

Also corrected was a typographical error for one of the instances
of the variable that defines the dependent libraries for the
test-Duct utility.