Update C++ unit tests to use dynamic OCALL management.
This update converts the following two C++ unit tests to use the
SRDEocall object to dynamically create the OCALL dispatch table:

test-fusion-c++

test-naaaim-c++

The test-naaaim-c++ utility was also updated to include unit test
seven that exercises the enclave instance of the RSAkey object.

The linkage directives for the test-naaaim-c++ utility was
modified since it was failing due to unresolved symbols from the
OpenSSL library in addition to the NAAAAIM library.