Finalize architecture for dynamic OCALL handling.
Work on attempting to add dynamic OCALL registration for the
SRDEnaaaim library indicated that a different strategy would be
needed.

It became apparent that having library specific methods in the
SRDEocall object turned out to be problematic.  This ended up
causing unresolved references when linking the test-fusion unit
test.

The final strategy chosen was to implement an array of relevant
OCALL's for each library.  This array is null-terminated which
makes reading the array straight forward.

The following method was added to the SRDEocall object:

->add_table

Which adds an array of function pointers to the current object.
This provides a consistent framework for all libraries and
applications to follow.

These changes allowed the OCALL infrastructure for the SRDEnaaaim
library to be integrated into an application.

A number of changes and fixes were also integrated, particularly
in the numbering of the slot positions for the SRDEfusion and
SRDEnaaaim enclave libraries.