Make OCALL table arguement type specific.
The ->boot_slot method call had taken a void * arguement type to
specify the OCALL dispatch table.  This had caused a number of
errors as the code was ported to use dynamic dispatch tables.

This commit changes the arguement type to a pointer to a struct
OCALL_api structure in order to make the call to the ->boot_slot
method more type specific.