Update quoting enclave ECALL interface.
Validation of the SRDE against the Intel version 2.6 prebuilt
libraries indicated there was a involving use of the quoting
enclave.

The Intel 2.5 release changed the API for the verify_blob() ECALL
to include a pointer to a data buffer that is loaded with the raw
CPU security version (CPUSVN).  This change appears to be
secondary to support for the Data Center Attestation Protocol
(DCAP).

The fix was straight forward and involved adding a pointer to the
structure that is passed to the ECALL.  That pointer is
configured to point to a 16 byte buffer allocated on the stack.

This API change was causing an enclave segmentation fault that was
picked up by the Attestation/test-remote-target unit test as well
as the test-Possum unit test.