Address SDK unit test build issues.
This update implements fixes for issues that arose as a result of
testing the SDK unit test sample code.

The install-dev target was not installing the OCALL header files for
the SRDEfusion and SRDEnaaaim libraries.  The SRDEfusion-ocall.h
and SRDEnaaaim-ocall.h header files are now installed as part of
the SDK.

The test-naaaim utility was immediately generating an enclave
segmentation fault.  This was traced to the fact that the enclave
linker flags were not specifying that the OpenSSL trusted
interface library (sgx_tsgxssl) be linked into the enclave in its
entirety.

The -Wl,--whole-archive and -Wl,--no-whole-archive clause was
modified to include this library along with the trusted runtime
library.