Implement RSAkey object in enclave context.
This commit adds an implementation of the RSAkey object to the
SRDEnaaaim library in order to support management of asymmetric
RSA keys in a trusted excecution environment.

In order to maintain API compatibility between trusted and
untrusted space this update also modifies the RSAkey methods in
the untrusted implementation.

Since loading keys via the OpenSSL file I/O strategy is not
currently available in enclave context the following two methods
were added to both the untrusted and trusted implementations:

->load_public

->load_private

These methods load either a public or private key from the
contents of a Buffer object.  Since a memory BIO is used for the
implementation the Buffer can contain a standard PEM encoded key.

The ->load_private method does not yet support an encrypted PEM
envelope.  It is assumed that enclave based encryption will be
used to encrypt/decrypt the contents of the Buffer.

In the enclave object implementation the ->load_public_key and
->load_private_key methods are reduced to implementations that
simply issue an error if they are invoked.