Implement additional SRDEfusion bridging code.
This commit implements additional shim code that allows
applications to be compiled without modification for an enclave
environment.

The most significant component of this update is to add support
for basic time management.  The code implementation is extremely
minimalistic at this point and essentially returns the time of
the start of the epoch.  The goal will be to provide additional
functionality in future updates that leverage a TPM to a sense of
monotonic time.

Also implemented in the shim code is an implementation of the
strptime() function.

The sprintf() function was also added despite its obvious danger
in order to increase compatibility between untrusted and enclave
based applications.