Implement allocation of KSS configuration elements.
One of the features that emerged the Intel 2.4 SDK was support
for Key Separation and Sharing (KSS).  Information on its
rationale is currently lacking but it would appear to be a method
for having an enclave signed by the same key generate different
keys based on a configuration identifiter and security version
value.

This update ads the config_id and config_svn structure elements
to structures which implement these elements.  The affected
structures are as follows:

SGX_secs

SGX_keyrequest

SGX_reportbody

The KSS security version (config_svn) is a component of a key
request and the SEALkey object that will be following explicitly
handles the field hence its incorporation even though the SRDE is
currently not supporting designation of a KSS configuration
identity.