Initial implementation of sealing object.
The set of changes in this commit implement the following
SRDEnaaaim object:

SEALkey

This object encapsulates all of the functionality needed to
generate and manage enclave sealing keys.  The current
implementation only implements support for MRSIGNER
based keys which are used to implement migratable keys.

A unit test was added to the test-naaaim unit to demonstrate that
the object will properly regenerate a key given a specific key
identifier value.