Implement IV and key accessor method.
Generation of the key does little good if there is no way to get
the keying material out of the object.  This update adds a
unified accessor named as follows:

->get_iv_key

That retrieves both the initialization vector and the key itself
from the object.

The unit test utility was modified to use this accessor method
after key re-generation in order to prove that the keying
elements retrieved are identical to those that were initially
generated.