Implement ASN1 encoding of key request.
This commit updates the encoding of the elements of the key
request into an ASN1 DER encoding bytestream.  The intent of this
is to provide a portable form of the components of a key request
that can be stored with the encoded blob in order to insure the
reproducibility of the keying material.

The following methods were added to the SEALkey object in order
to faciliate access to this functionality:

->get_request

->set_request

The enclave unit test was modified to demonstrate reproducibility
of the key using this infrastructure.