Rationalize data alignments.
The SGX instructions that request reports and keys have strict
alignment constraints on their operands.  The code had previously
not added explicit alignment constraints for the target
information and keyrequest structures.  This update adds explicit
alignment directives for these structures.

The formatting of the alignment directives was normalized so that
the alignment directive are included between the variable type
and the variable name itself.