Isolate key generation in private method.
In order to avoid code replication the mechanics of generating
and deriving an initialization vector and key were isolated into
an internal method.  This mode takes as an arguement the key
policy to be used.

This sets the stage for adding a method to generate an enclave
(MRENCLAVE) based signing key.