Implement generation of MRenclave based keys.
With the consolidation of the key generation into an independent
method it was a straight forward to implement the generation of
keys based on the measurement of the enclave.  The ->generate_mrenclave
method implement to invoke this type of key generation.

It should be noted that this type of key generation is extremely
fragile compared to keys that are derived from the signing
identity of the enclave.  As a result this type of key is most
useful within the context of a given enclave.