Implement key shrouding.
The threat of side-channel attacks against the contents of an
enclave, remote as it is on properly updated platforms, suggests
the need to protect the values of keys that are generated inside
of an enclave by the ENCLU[EGETKEY] instruction.  This update
implements shrouding of a key generated by the SEALkey object
with an MRENCLAVE based key.

After the desired key and initialization vector are generated an
MRENCLAVE based key is generated and is XOR'ed with the key that
is stored in the object state structure.  The shrouding key is
re-generated in the ->get_iv_key method and is applied to the
Buffer object containing the key that is held by the caller.

This model operates under the assumption that the key will be
quickly consumed and then disposed of to avoid it being held in
memory by the code that has requested key generation.