Implement generation of static keys.
This update adds the following method to the SEALkey object:

->generate_static_key

The purpose of this method is to implement the generation of a
static key.  This key is generated with a fixed key identifier
provided by the caller and has no dependency on any of the
security version components supported by the SGX_keyrequest
structure.

A static key can be generated by any subsequent enclaves and is
not affected by microcode updates.  The only dependency is the
key identifier value supplied by the caller.

The test-naaaim unit test enclave was updated to generate a
static key using a key identifier file that is fixed in the unit
test enclave.