Use correct include file.
Due to a long standing cut and paste error the source for the
test-naaaim unit test enclave was including the wrong header
file.  This did not have an operative impact on the unit test but
for correctness purposes, since the unit test is distributed with
the SDK, the correct header file was included.