Fix compilation of SmartCard object.
The file containing the code to implement the SmartCard object
was not being successfully compiled on RHEL systems with the
SRDE_libs installation.  This was secondary to the include file
directory path not being properly set when the non musl-GCC
compiler was being invoked.

The target to create the SmartCard.o file was also incorrectly
specified and the Makefile was attempting to build a binary
rather then an object file.

The Makefile conditional that was selecting for the correct
location for the OpenSSL include files and libraries was modified
to set the SC_INCLUDE variable to point to the directory
containing the include files for the pcsc-lite implementation for
the requested compiler.

The target to create the object file was also properly defined.