Implement SEALEDblob object.
This update implements support for an object that encapsulates
the enclave based sealing of an arbitrary data 'blob'.  The blob
is provided in the form of a Buffer object.

The sealed data consistes of two components.  The first is an
SGX_keyrequest structure that defines the key components needed
to create the derived key that is used to encrypt and integrity
protect the primary  data to be protected.  This encrypted and
integrity protected block of primary data is the second of the
two components.

A static enclave key is used to encrypt and integrity protect the
keying component data block.  While not strictly necessary it does
significantly increase the challenge associated with the
potential elucidation of any information which could be used to
attack the nature of the primary key.

Both data blocks are confidentiality protected with AES256.
Integrity of the encrypted data block is implemented with a
SHA256_hmac checksum.  The key used for the HMAC checksum is as
follows:

HKEY = SHA256(IV || KEY)

Where IV is the initialization vector, KEY is the encryption
key and || is the concantenation operator.

The checksum for the encrypted keying element structure is over
only that data block.  The checksum created with the primary
data element key is generated over the entire data block that
includes the encrypted and integrity protected keying element
blob.