Fix key policy designation constants.
The values for constants that define either enclave or signer key
derivation policies had somehow become inverted.  As a result
specifying key derivation based on signer identity was resulting
in a sealing key that would change whenever the measurement value
of the enclave changed.

The SRDE master include file was updated to correct this issue
and add all of the key policy constants that are currently
defined by the Intel SDK.

As a continuation of our policy of removing references to SGX the
names of the contents were changed from the following prefix:

SGX_

To:

SRDE_

The changes were propagated throughout the source tree.  The
current code is now verified to create enclaves that generate
consistent and invariant keys when an SRDE_KEYPOLICY_SIGNER
derivation policy is specified.