Do not build sdk-compute-size utility by default.
The sdk-compute-size utility was implemented during the initial
development of the SRDE when it was necessary to debug problems
associated with equivalency of structure sizes between the two
runtimes.  The utility requires access to the internal Intel SDK
headers in order to compuete these sizes.

The utility is of no general use to the SRDE tooling and requires
access to the Intel SDK in possibly a version dependent fashion.

This update removes the utility from the TOOLS variable in the
SRDE/Makefile so that it is not built by default.  Building it
can be achieved by issueing the following command in the top of
the SRDE directory heirarchy:

make sdk-compute-size

If access to the utility is needed.