Make RSA public key loading compatible.
The following RSAkey methods:

->get_public

->load_public

Were using the the following methods to read and save RSA keys:

PEM_write_bio_RSAPublicKey

PEM_read_bio_RSAPublicKey

Which uses a PEM encoding format that is incompatible with the
->get_public_key method.  The PEM_{read,write}_bio_RSAPublicKey
routines are also incompatible with the way that openssl handles
public keys in its command line interface.

The update replaces the above two routines with the following
routines:

write_bio_RSA_PUBKEY

read_bio_RSA_PUBKEY

Which causes the public key to be encoded in a manner that is
compatible with the openssl binary as well as the file based
methods the RSAkey object uses for loading keys.