Capture attestation report headers.
A critical element to the ability of a verifying party, the
entity that evaluates a remote attestation report, is to check
the Intel supplied signature that is supplied over the
attestation report.  The signature and the certificate that
was used to generate the signature is provided by the following
two headers in the HTTPS response:

x-iasreport-signature:

x-iasreport-signing-certificate:

This update modifies the arguements that are passed to the wget
command in order to specify that the response headers be included
in the body of the response.  In addition the command-line
arguements that had previously generated a logfile were removed
in order to prevent creation of the file.

The arguements to run in verbose mode and show the server
responses were also removed.