Extract IAS certificate and signature from attestation report.
The Intel Authentication Services (IAS) report is returned with a
signature over the enclave attestation report.  The signature is
generated against the IAS signing certificate.  That certificate
in turn is signed by the IAS Certificate Authority certificate.

The signature and the certificate stack (two certificates) are
returned as headers on the REST response.

This update provides infrastructure that abstracts the signatures
and the verifying certificates as separate entities that are
maintained in the SRDEquote object state.  In addition the report
body proper is abstracted and stored as an independent data
object in the object state so that its signature can be verified.

At this point in the source tree the signature and the
certificate stack are printed.  Subsequent updates will add
verification of the report signature.