Implement certificate based public key loading.
One of the most important appliations of asymmetric keys are
signature verification.  One of the common sources for a signing
key is an X509 certificate.  This commit implements basic support
for setting the public key mode of the RSAkey object by loading a
certificate and setting the public key to the public key that is
embedded in the certificate.

In order to properly support this the RSAkey object state was
updated to include an X509 certificate object.  Testing revealed
the freeing the object created problems with the abstracted RSA
key even when a reference to the key structure was held.