Add support for signature verification.
This commit adds the following method to the RSAkey objects in
both untrusted and enclave mode:

->verify

This method builds on top of the ->load_certificate method that
was previously committed to enable the ability to verify a
signature that was made with the private key tied to the loaded
certificate.

The RSAkey_test utility was updated to exercise this
functionality.  The test-naaaim unit test enclave was also
updated to verify a signature.

In the case of the enclave based signature verification this
update adds the private key, certificate and signature in array
format to the enclave.  This provides a framework for the unit
test code to be self-standing.