Implement support for IAS report validation.
This commit adds the following method to the standard and enclave
versions of the SRDEquote object:

->validate_report

This method validates a previously decoded IAS report.  The
validation status of the report is returned via a pointer to a
boolean variable maintained in the caller.

While the ->validate_report method is designed to be a standalone
method the ->dump_report method was modified to print out the
validation status of the report as well.

In addition to validating the report signature the
->validate_report method validates the signing certificate that
is supplied in response to the attestation report request.  A
copy of the Intel IAS certificate CA is embedded in the SRDEquote
objects.  This certificate is used to validate the signature
certificate provided in the REST response.

The CA certificate is returned in addition to the signing
certificate but embedding of the CA certificate in the SRDE
provides an additional confirmation as to the integrity of the
signing certificate.