Implement public key signature verification.
This commit modifies both the standard and enclave version of the
RSAkey object to allow signature verification based on a public
key rather then just a certificate.

To support this a 'certificate' object type enumeration was
defined to distinguish the key from when a certificate is used as
the source of a public key rather then an explicitly loaded
public key.

The ->verify method loads the correct key into the locally scoped
instance of an EVP_PKEY object that is used to verify the
signature based on the enumeration type of the object.  The
loading of the object state instance of the RSA key is removed
from the ->load_certificate method.

Both the trusted and enclave unit tests were updated to
demonstrate the ability of the RSAkey object to verify a
signature when provided a public key.  This required the
embedding of the public key into the the unit test code.