Add signature generation support to RSAkey object.
This update implements the ability for the standard and enclave
versions of the RSAkey object to generate signatures based on a
private key in the possession of the object.

Signature generation is implemented via the following method:

->sign

Signature generation is compatible with the ->verify method that
has been previously implemented.

The RSAkey_test unit test was updated to add support for
generating a signature with the 'sign' verb.  The verb requires the
name of a file containing a private key and the name of a file
containing the data over which the signature is to be generated.

By default the signature is printed out in Buffer->hprint form.
If an additional filename is added to the unit test directive the
signature will be written to a file by that name.  This allows
the signature to be fed back into the 'verify' functionality of
the unit test.

The enclave based unit test was updated to implement the
generation and verification of a signature using the private key
that is embedded in the test-naaaim-enclave.c file.