Implement passphrase object.
This update implements the Prompt object which encapsulates the
functionality needed to prompt the user for a passphrase using
the OpenSSL UI API.

Currently all of the functionality is implemented in the ->get
method.

This object supports simple prompting for a passphrase in
addition to verifying a passphrase that has been entered.
Verification is requested by passing a non-NULL String object
that contains a preamble phrase to be displayed before the phrase
that was used to solicit the initial passphrase.