Flush stand output on each SRDEfusion print OCALL.
When all of the SRDEfusion OCALL handling was consolidated into a
single implementation the fflush(stdout) call that some
applications had implemented was not transferred into the
canonical implementation.

As a result applications such as cboot-mgr were not performing
in an expected fashion.  Most importantly the application would
not give any indication of a successful start.

This update simply adds the fflush() function call to the
canonical implementation.  The performance impacts of this are
trivial in comparison to the OCCAL exit/re-entry costs.